Merge safeAreaContext into main

+++ src/screens/SettingsScreen.js
@@ -1,4 +1,5 @@
-import { StyleSheet, Text, View, SafeAreaView, Button, Switch, Pressable } from 'react-native';
+import { StyleSheet, Text, View, Button, Switch, Pressable } from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
 import { useState, useEffect } from 'react';
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import {styles} from '../styles';
@@ -52,7 +53,14 @@ export default function SettingsScreen() {
 
         try {
             const audioDirectory = FileSystem.documentDirectory + 'audio/';
+            const directoryInfo = await FileSystem.getInfoAsync(audioDirectory);
+            console.log(directoryInfo);
+            if(directoryInfo.exists == false) {
+                console.log("directory does not exist - cancelling deletion as it is unneccesary.");
+                return;
+            }
             const audioFiles = await FileSystem.readDirectoryAsync(audioDirectory);
+            if (audioFiles == null) return;
             await Promise.all(audioFiles.map((fileName) => {
                 FileSystem.deleteAsync(audioDirectory + fileName, { idempotent: true })
             }));
